refactor: create plugin folder
