+++ node-pubchem/src/src/server/index.js
@@ -3,6 +3,8 @@ import { join } from 'path';
 
 import fastifySwagger from 'fastify-swagger';
 
+import { recursiveDir } from './recursiveDir.js';
+
 export default async function setupV1(app, _, done) {
   app.register(fastifySwagger, {
     routePrefix: '/documentation',
@@ -20,19 +22,20 @@ export default async function setupV1(app, _, done) {
     reply.redirect('/v1/documentation');
   });
 
-  const url = new URL('routes', import.meta.url);
-  const folders = await readdir(url);
-  for (let folder of folders) {
-    const folderURL = new URL(join('routes', folder), url);
-    for (let routeName of await readdir(folderURL)) {
-      const routeURL = new URL(join('routes', folder, routeName), url);
-      const stats = await lstat(routeURL);
-      if (!stats.isFile()) continue;
+  const url = new URL('../plugins/', import.meta.url);
+  const routeURLs = (await recursiveDir(url)).filter((file) =>
+    file.href.match(/routes/),
+  );
 
-      const route = (await import(routeURL)).default;
-      if (typeof route.schema !== 'object') continue;
-      app.route(route);
-    }
+  for (let routeURL of routeURLs) {
+    const route = (await import(routeURL)).default;
+    if (typeof route.schema !== 'object') continue;
+    const path = routeURL.pathname
+      .replace(url.pathname, '')
+      .replace('/routes', '')
+      .replace('.js', '');
+    route.url = path;
+    app.route(route);
   }
 
   done();
