+++ node-pubchem/src/src/server/utils/getFields.js
@@ -2,7 +2,7 @@ import Debug from 'debug';
 
 const debug = Debug('getFields');
 
-function getFields(fields = '') {
+export function getFields(fields = '') {
   let result = { id: '$_id' };
   for (let field of fields.split(',').filter((field) => field)) {
     result[field] = 1;
@@ -10,5 +10,3 @@ function getFields(fields = '') {
   debug(`List of required fields: ${JSON.stringify(result)}`);
   return result;
 }
-
-export default getFields;
